Separated Integration (DefaultProvider) tests from Core tests for Parser.
