+++ PDM.sln
@@ -17,6 +17,14 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PDM.Parser", "src\PDM.Parser\PDM.Parser.csproj", "{C0C82F43-37E8-4BEF-ACBD-4D31D41A8E22}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{460F91EE-7A06-4B59-AE80-9290FED6C66D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tst", "tst", "{7214354A-8565-48C9-95F0-C18112DA8EB3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PDM.DefaultProviders.Test", "tst\PDM.DefaultProviders.Test\PDM.DefaultProviders.Test.csproj", "{30A85217-BD0C-4E01-A439-083CD2BD51DB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PDM.TestUtils", "tst\PDM.TestUtils\PDM.TestUtils.csproj", "{DA1846FB-7690-4F8F-8FCE-E0C87BC77652}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,9 +43,24 @@ Global
 		{C0C82F43-37E8-4BEF-ACBD-4D31D41A8E22}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C0C82F43-37E8-4BEF-ACBD-4D31D41A8E22}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C0C82F43-37E8-4BEF-ACBD-4D31D41A8E22}.Release|Any CPU.Build.0 = Release|Any CPU
+		{30A85217-BD0C-4E01-A439-083CD2BD51DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{30A85217-BD0C-4E01-A439-083CD2BD51DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{30A85217-BD0C-4E01-A439-083CD2BD51DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{30A85217-BD0C-4E01-A439-083CD2BD51DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DA1846FB-7690-4F8F-8FCE-E0C87BC77652}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DA1846FB-7690-4F8F-8FCE-E0C87BC77652}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DA1846FB-7690-4F8F-8FCE-E0C87BC77652}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DA1846FB-7690-4F8F-8FCE-E0C87BC77652}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{76C2AB7D-12E6-40D6-A87E-71FF2A8F23D7} = {460F91EE-7A06-4B59-AE80-9290FED6C66D}
+		{AF787EE4-BC92-421B-A90D-0E960E506FCD} = {7214354A-8565-48C9-95F0-C18112DA8EB3}
+		{C0C82F43-37E8-4BEF-ACBD-4D31D41A8E22} = {460F91EE-7A06-4B59-AE80-9290FED6C66D}
+		{30A85217-BD0C-4E01-A439-083CD2BD51DB} = {7214354A-8565-48C9-95F0-C18112DA8EB3}
+		{DA1846FB-7690-4F8F-8FCE-E0C87BC77652} = {7214354A-8565-48C9-95F0-C18112DA8EB3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {23D296E3-1C54-46AF-9567-DDE6D118674B}
