+++ src/PDM.Core/Entities/Varint.cs
@@ -1,4 +1,6 @@
-﻿namespace PDM.Entities;
+﻿using System.Globalization;
+
+namespace PDM.Entities;
 
 public sealed record Varint
 {
@@ -6,21 +8,20 @@ public sealed record Varint
 
     public int WireLength => this.RawData.Length;
 
-    internal byte[] DecodedData
-        => this.RawData
-        .Select(d => Convert.ToByte(d & 127))
-        .ToArray();
-
 #pragma warning disable CS3003 // Type is not CLS-compliant
     public UInt64 Value => CalculateValue(this.DecodedData);
 #pragma warning restore CS3003 // Type is not CLS-compliant
 
 
-    internal Varint(byte[] rawData) 
+    internal Varint(byte[] rawData)
         => this.RawData = rawData;
 
-    internal Varint(long value)
-        :this((ulong)value)
+    public Varint(long value)
+        : this((ulong)value)
+    { }
+
+    public Varint(double value)
+        : this((ulong)value)
     { }
 
     internal Varint(ulong value)
@@ -59,6 +60,37 @@ public sealed record Varint
         return new Varint(message[0..length]);
     }
 
+    public static Varint Create(object value)
+    {
+        return value switch
+        {
+            long => new Varint((long)value),
+            ulong => new Varint((ulong)value),
+            double => new Varint((double)value),
+            string => Varint.CreateFromString((string)value ?? string.Empty),
+            _ => throw new NotImplementedException($"{value} ({value.GetType().Name}) cannot be parsed as a numeric")
+        };
+    }
+
+    private static Varint CreateFromString(string value)
+    {
+        Varint result;
+        if (long.TryParse(value.ToString(), NumberStyles.Integer, CultureInfo.InvariantCulture, out var longValue))
+            result = new Varint(longValue);
+        else if (ulong.TryParse(value.ToString(), NumberStyles.Integer, CultureInfo.InvariantCulture, out var ulongValue))
+            result = new Varint(ulongValue);
+        else if (double.TryParse(value.ToString(), NumberStyles.Float, CultureInfo.InvariantCulture, out var doubleValue))
+            result = new Varint(doubleValue);
+        else
+            throw new NotImplementedException("{value} cannot be parsed as a numeric");
+        return result;
+    }
+
+    private byte[] DecodedData
+        => this.RawData
+        .Select(d => Convert.ToByte(d & 127))
+        .ToArray();
+
     private static UInt64 CalculateValue(byte[] decodedBytes)
     {
         UInt64 total = 0;
@@ -67,7 +99,5 @@ public sealed record Varint
         return total;
     }
 
-    //public override string ToString()
-    //    => this.Value.ToString(CultureInfo.CurrentCulture);
 }
