+++ tst/PDM.Core.Test/PDM.Core.Test.csproj
@@ -8,31 +8,12 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<None Remove="ProtoBuf\AllTypes.proto" />
-		<None Remove="ProtoBuf\MismatchedType.proto" />
-		<None Remove="ProtoBuf\WeirdnessDemo.proto" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Protobuf Include="ProtoBuf\MismatchedType.proto" GrpcServices="Client" />
-		<Protobuf Include="ProtoBuf\AllTypes.proto" GrpcServices="Client">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</Protobuf>
-		<Protobuf Include="ProtoBuf\TwoFields.proto" GrpcServices="Client">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</Protobuf>
-		<Protobuf Include="ProtoBuf\ThreeFields.proto" GrpcServices="Client">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</Protobuf>
-		<Protobuf Include="ProtoBuf\WeirdnessDemo.proto" GrpcServices="Client" />
-	</ItemGroup>
-
-	<ItemGroup>
 		<PackageReference Include="Grpc.Tools" Version="2.51.0" />
-		<PackageReference Include="Google.Protobuf" Version="3.22.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="6.0.0" />
+		<PackageReference Include="Google.Protobuf" Version="3.22.1" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="6.0.0" />
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+		<PackageReference Include="Moq" Version="4.18.4" />
 		<PackageReference Include="Serilog" Version="2.12.0" />
 		<PackageReference Include="Serilog.Sinks.Xunit2" Version="1.1.0" />
 		<PackageReference Include="Serilog.Extensions.Hosting" Version="5.0.1" />
@@ -50,7 +31,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\..\src\PDM.Core\PDM.Core.csproj" />
-		<ProjectReference Include="..\..\src\PDM.Parser\PDM.Parser.csproj" />
+		<ProjectReference Include="..\PDM.TestUtils\PDM.TestUtils.csproj" />
 	</ItemGroup>
 
 </Project>
