+++ tst/PDM.DefaultProviders.Test/ProtobufMapper_MapAsync_HasSomeWeirdness.cs
@@ -1,12 +1,12 @@
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using PDM.Builders;
-using PDM.Core.Test.Extensions;
 using PDM.Parser.Extensions;
 using Serilog;
 using Xunit.Abstractions;
+using PDM.TestUtils.Extensions;
+using PDM.TestUtils.ProtoBuf;
 
-namespace PDM.Core.Test;
+namespace PDM.DefaultProviders.Test;
 
 [ExcludeFromCodeCoverage]
 [Collection("MapperTests")]
@@ -74,7 +74,7 @@ public class ProtobufMapper_MapAsync_HasSomeWeirdness
     public async Task EncodeSignedIntsDifferently()
     {
         var expected = 2132; // Zig-Zag encoded 1066
-        var sourceData = new ProtoBuf.AllTypes()
+        var sourceData = new TestUtils.ProtoBuf.AllTypes()
         {
             SInt32Value = 1066
         };
@@ -88,7 +88,7 @@ public class ProtobufMapper_MapAsync_HasSomeWeirdness
         var target = _serviceProvider.GetMapper(targetMapping);
         var actual = await target.MapAsync(sourceMessage); ;
 
-        var actualData = ProtoBuf.AllTypes.Parser.ParseFrom(actual);
+        var actualData = AllTypes.Parser.ParseFrom(actual);
 
         // Since the value was encoded using Zig-Zag encoding
         // we won't get the "intended" value of 1066, but
