+++ tst/PDM.TestUtils/Extensions/MessageExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Google.Protobuf;
 
-namespace PDM.Core.Test.Extensions;
+namespace PDM.TestUtils.Extensions;
 
 [ExcludeFromCodeCoverage]
 public static class MessageExtensions
