+++ tst/PDM.TestUtils/PDM.TestUtils.csproj
@@ -1,0 +1,39 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net6.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<None Remove="ProtoBuf\MismatchedType.proto" />
+		<None Remove="ProtoBuf\ThreeFields.proto" />
+		<None Remove="ProtoBuf\TwoFields.proto" />
+		<None Remove="ProtoBuf\WeirdnessDemo.proto" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
+		<PackageReference Include="Grpc.Tools" Version="2.52.0">
+		  <PrivateAssets>all</PrivateAssets>
+		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="Google.Protobuf" Version="3.22.1" />
+		<PackageReference Include="TestHelperExtensions" Version="2.1.0.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\src\PDM.Core\PDM.Core.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Protobuf Include="ProtoBuf\AllTypes.proto" GrpcServices="Client" />
+		<Protobuf Include="ProtoBuf\MismatchedType.proto" GrpcServices="Client" />
+		<Protobuf Include="ProtoBuf\ThreeFields.proto" GrpcServices="Client" />
+		<Protobuf Include="ProtoBuf\TwoFields.proto" GrpcServices="Client" />
+		<Protobuf Include="ProtoBuf\WeirdnessDemo.proto" GrpcServices="Client" />
+	</ItemGroup>
+
+</Project>
